Use Kotlin/Native's built-in NSMutableArray wrapping
I don't believe this is safe to do as the type is not public nor written into caller binaries anywhere. The C++ APIs seem to have to be stable, but I can't convince myself that it also makes the Kotlin type stable.

They also don't implement RandomAccess, although I sent a PR for that.